refactor: 타입/상수 분리 및 단일 진입점 패턴으로 개선
- 도메인 모델(compare.ts)과 비교 결과(compare-result.ts) 타입 분리
- 상수를 constants/compare.ts로 분리
- features/compare 에서 단일 진입점 사용하도록 import문 수정
