+++ apps/knockdog/src/entities/compare/index.ts
@@ -1,11 +1,47 @@
-// lib
-export * from './lib/serialize';
-export * from './lib/utils';
+/** constants */
+export {
+  CTAG_MAP,
+  PRODUCT_TYPE,
+  TRANSPORTATION_TYPE,
+  REFERENCE_POINT_TYPE,
+  DAY_OF_WEEK_SHORT,
+  CLOSED_DAYS,
+} from './model/constants/compare';
 
-// model
-export * from './model/types';
+/** config */
+export { createComparisonsQueryOptions } from './config/comparisonsQueryKeys';
 
-// ui
+/** lib */
+export { parseTimeStrToMinutes, parseMinutesToTimeStr, getClosedDaysText } from './lib/formatters';
+export {
+  resolveIds,
+  s3ToUrl,
+  mapToSimpleItem,
+  getProduct,
+  getProductMin,
+  getProductMax,
+  getTransitTime,
+  getDistanceString,
+} from './lib/utils';
+export { serializeCategories } from './lib/serialize';
+
+/** model */
+export type {
+  KindergartenComparison,
+  CTag,
+  ProductType,
+  ReferencePointType,
+  TransportationType,
+  DayOfWeekShort,
+  SimpleComparisonItem,
+  DetailComparisonItem,
+  PriceDetailComparison,
+  DistanceComparisonsByRef,
+  DistanceComparisonsByTransport,
+  DistanceDetailComparison,
+} from './model/types';
+
+/** ui */
 export { Summary } from './ui/Summary';
 export { Description } from './ui/Description';
 export { Detail } from './ui/Detail';
