+++ apps/knockdog/src/entities/compare/lib/utils.ts
@@ -1,15 +1,16 @@
-import {
-  CLOSED_DAYS,
-  KindergartenComparison,
-  ProductType,
-  SimpleComparisonItem,
-  TransportationType,
-} from '../model/types';
+import type { KindergartenComparison, ProductType, TransportationType } from '../model/compare';
+import type { SimpleComparisonItem } from '../model/compare-result';
 
-export function resolveIds(searchParams: URLSearchParams): string[] {
+/**
+ * URLSearchParams에서 유치원 ID 목록을 추출합니다.
+ * @param searchParams URLSearchParams 객체
+ * @returns 유치원 ID 배열
+ */
+function resolveIds(searchParams: URLSearchParams): string[] {
   // ids=aaa&ids=bbb
   const repeated = searchParams.getAll('ids').filter(Boolean);
   if (repeated.length >= 2) return repeated;
+
   // ids=aaa,bbb
   const commaJoined = searchParams.get('ids');
   if (commaJoined) {
@@ -21,83 +22,54 @@ export function resolveIds(searchParams: URLSearchParams): string[] {
   return [];
 }
 
-export function s3ToUrl(s3Key?: string) {
+function s3ToUrl(s3Key?: string) {
   if (!s3Key) return undefined;
   const CDN = process.env.NEXT_PUBLIC_IMAGE_BASE_URL;
   return CDN ? `${CDN}${encodeURI(s3Key)}` : undefined;
 }
 
-/* =========================
- * SUMMARY CALCULATION
- * ========================= */
-
-// 시간 문자열을 분으로 변환 ("2시간 49분" -> 169)
-export function parseTimeStrToMinutes(timeStr: string): number {
-  if (!timeStr) return 0;
-
-  let totalMinutes = 0;
-  const hourMatch = timeStr.match(/(\d+)시간/);
-  const minMatch = timeStr.match(/(\d+)분/);
-
-  if (hourMatch) totalMinutes += parseInt(hourMatch[1] || '0') * 60;
-  if (minMatch) totalMinutes += parseInt(minMatch[1] || '0');
-
-  return totalMinutes;
-}
-
-// 분 단위를 시간 문자열로 변환 (169 -> "2시간 49분")
-export function parseMinutesToTimeStr(minutes: number): string {
-  if (!minutes) {
-    return '-';
-  }
-
-  const hours = Math.floor(minutes / 60);
-  const mins = minutes % 60;
-
-  const parts: string[] = [];
-  if (hours > 0) {
-    parts.push(`${hours}시간`);
-  }
-  if (mins > 0) {
-    parts.push(`${mins}분`);
-  }
-
-  return parts.join(' ');
-}
-
-export function mapToSimpleItem(kg: KindergartenComparison): SimpleComparisonItem {
+function mapToSimpleItem(kg: KindergartenComparison): SimpleComparisonItem {
   return {
     name: kg.name,
     avatar: s3ToUrl(kg.thumbnailS3Key) ?? '',
   };
 }
 
-/* =========================
- * DETAILS CALCULATION
- * ========================= */
+function getProduct(kg?: KindergartenComparison | null, prodType?: ProductType | null) {
+  return kg?.pricing?.products?.find((product) => product?.productType === prodType) ?? null;
+}
 
-type ProdKey = ProductType;
+function getProductMin(kg?: KindergartenComparison | null, prodType?: ProductType | null) {
+  return getProduct(kg, prodType)?.min?.price ?? null;
+}
 
-export const getProduct = (kg?: KindergartenComparison | null, prodType?: ProdKey | null) =>
-  kg?.pricing?.products?.find((product) => product?.productType === prodType) ?? null;
+function getProductMax(kg?: KindergartenComparison | null, prodType?: ProductType | null) {
+  return getProduct(kg, prodType)?.max?.price ?? null;
+}
 
-export const getProductMin = (kg?: KindergartenComparison | null, prodType?: ProdKey | null) =>
-  getProduct(kg, prodType)?.min?.price ?? null;
-
-export const getProductMax = (kg?: KindergartenComparison | null, prodType?: ProdKey | null) =>
-  getProduct(kg, prodType)?.max?.price ?? null;
-
-export const getTransitTime = (
+function getTransitTime(
   kg?: KindergartenComparison | null,
   refPoint: string = 'HOME',
   mode: TransportationType = 'WALKING'
-) =>
-  kg?.distance
-    ?.find((distance) => distance?.referencePoint === refPoint)
-    ?.transitTimes?.find((transit) => transit?.type === mode)?.time ?? '-';
+) {
+  return (
+    kg?.distance
+      ?.find((distance) => distance?.referencePoint === refPoint)
+      ?.transitTimes?.find((transit) => transit?.type === mode)?.time ?? '-'
+  );
+}
 
-export const getDistanceString = (kg?: KindergartenComparison | null, refPoint: string = 'HOME') =>
-  kg?.distance?.find((distance) => distance?.referencePoint === refPoint)?.distance ?? '-';
+function getDistanceString(kg?: KindergartenComparison | null, refPoint: string = 'HOME') {
+  return kg?.distance?.find((distance) => distance?.referencePoint === refPoint)?.distance ?? '-';
+}
 
-export const getClosedDaysText = (kg?: KindergartenComparison | null) =>
-  (kg?.operatingSchedule?.closedDays ?? []).map((dayKey) => CLOSED_DAYS[dayKey]).join(', ') || '-';
+export {
+  resolveIds,
+  s3ToUrl,
+  mapToSimpleItem,
+  getProduct,
+  getProductMin,
+  getProductMax,
+  getTransitTime,
+  getDistanceString,
+};
