+++ apps/knockdog/src/features/compare/api/useComparisonsQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
-import { createComparisonsQueryOptions } from '@entities/compare/config/comparisonsQueryKeys';
-import { KindergartenComparison } from '@entities/compare/model/types';
+import type { KindergartenComparison } from '@entities/compare';
+import { createComparisonsQueryOptions } from '@entities/compare';
 import { ApiResponse } from '@shared/api';
 
 function useComparisonsQuery(ids: string[]) {
