+++ apps/knockdog/src/features/compare/lib/findShortestFromComparisons.ts
@@ -1,4 +1,4 @@
-import {
+import type {
   ReferencePointType,
   TransportationType,
   DistanceComparisonsByRef,
