+++ apps/knockdog/src/shared/utils/index.ts
@@ -1,4 +1,4 @@
-export { ellipsisText } from './text';
+export { ellipsisText, getSubjectParticle, getDirectionParticle } from './text';
 export { tokenUtils } from './token';
 export { isNative } from './isNative';
 export { eventBus } from './eventBus';
